# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ocean Weligama API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/bookings/check:
    post:
      operationId: checkAvailabilityAndPrice
      tags: [bookings]
      summary: Check availability and get price quote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingCheck"
      responses:
        "200":
          description: Price quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingQuote"
  /v1/bookings:
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Validation error or unavailable
  /v1/bookings/{reference}:
    get:
      operationId: getBookingByReference
      tags: [bookings]
      summary: Get booking by reference number (public confirmation)
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Booking not found

  # PUBLIC REVIEWS
...
  /v1/contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit a contact message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Message submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResult"

  # ANALYTICS
  /v1/analytics/track:
    post:
      operationId: trackPageView
      tags: [analytics]
      summary: Track a page view (cookieless)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PageViewInput"
      responses:
        "200":
          description: Tracked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackResult"

  # ADMIN AUTH
...
  /v1/admin/rooms:
    get:
      operationId: adminListRooms
      tags: [admin-rooms]
      summary: List all rooms including inactive
      responses:
        "200":
          description: All rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
    post:
      operationId: adminCreateRoom
      tags: [admin-rooms]
      summary: Create new room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomInput"
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /v1/admin/rooms/{id}:
    patch:
      operationId: adminUpdateRoom
      tags: [admin-rooms]
      summary: Update room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomUpdate"
      responses:
        "200":
          description: Updated room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
    delete:
      operationId: adminDeleteRoom
      tags: [admin-rooms]
      summary: Soft delete room
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted

  # ADMIN SERVICES
...
  /v1/blogs:
    get:
      operationId: listBlogs
      tags: [blogs]
      summary: List active blogs
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of blogs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Blog"
  /v1/blogs/{id}:
    get:
      operationId: getBlog
      tags: [blogs]
      summary: Get blog detail by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blog detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Blog"

  # ADMIN BLOGS
...
components:
  schemas:
...
    Blog:
      type: object
      required: [id, title, description, image, category, date, createdAt, updatedAt]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        image:
          type: string
        category:
          type: string
        date:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
...
    Room:
      type: object
      required: [id, slug, type, status, maxGuests, bedrooms, bathrooms, basePricePerNight, currency, name, description]
      properties:
        id:
          type: string
        slug:
          type: string
        type:
          type: string
          enum: [room, villa, dormitory, suite]
        status:
          type: string
          enum: [active, maintenance, hidden]
        maxGuests:
          type: integer
        bedrooms:
          type: integer
        bathrooms:
          type: integer
        sizeM2:
          type: ["integer", "null"]
        basePricePerNight:
          type: string
        cleaningFee:
          type: ["string", "null"]
        currency:
          type: string
        isFeatured:
          type: boolean
        heroImageUrl:
          type: ["string", "null"]
        name:
          type: string
        description:
          type: string
        shortDesc:
          type: ["string", "null"]
        amenities:
          type: array
          items:
            type: string
        category:
          type: string
        gallery:
          type: ["array", "null"]
          items:
            type: string
        highlights:
          type: ["array", "null"]
          items:
            type: string
...
    RoomInput:
      type: object
      required: [slug, type, maxGuests, bedrooms, bathrooms, basePricePerNight, currency, name, description]
      properties:
        slug:
          type: string
        type:
          type: string
          enum: [room, villa, dormitory, suite]
        maxGuests:
          type: integer
        bedrooms:
          type: integer
        bathrooms:
          type: integer
        sizeM2:
          type: ["integer", "null"]
        basePricePerNight:
          type: string
        cleaningFee:
          type: ["string", "null"]
        currency:
          type: string
        isFeatured:
          type: boolean
        heroImageUrl:
          type: ["string", "null"]
        name:
          type: string
        description:
          type: string
        category:
          type: ["string", "null"]
        gallery:
          type: ["array", "null"]
          items:
            type: string
        highlights:
          type: ["array", "null"]
          items:
            type: string
    RoomUpdate:
      type: object
      properties:
        type:
          type: string
        status:
          type: string
        maxGuests:
          type: integer
        basePricePerNight:
          type: string
        heroImageUrl:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        name:
          type: string
        description:
          type: string
        category:
          type: ["string", "null"]
        gallery:
          type: ["array", "null"]
          items:
            type: string
        highlights:
          type: ["array", "null"]
          items:
            type: string
...
    Booking:
      type: object
      required: [id, reference, roomId, checkIn, checkOut, nights, guestCount, guestFullName, guestEmail, guestPhone, totalAmount, currency, status, paymentStatus, paymentMethod, createdAt]
      properties:
        id:
          type: string
        reference:
          type: string
        roomId:
          type: string
        roomName:
          type: ["string", "null"]
        checkIn:
          type: string
        checkOut:
          type: string
        nights:
          type: integer
        guestCount:
          type: integer
        guestFullName:
          type: string
        guestEmail:
          type: string
        guestPhone:
          type: string
        guestNationality:
          type: ["string", "null"]
        guestCountryCode:
          type: ["string", "null"]
        specialRequests:
          type: ["string", "null"]
        roomRatePerNight:
          type: string
        roomSubtotal:
          type: string
        servicesSubtotal:
          type: string
        cleaningFee:
          type: string
        totalAmount:
          type: string
        currency:
          type: string
        depositAmount:
          type: string
        status:
          type: string
          enum: [pending, confirmed, checked_in, checked_out, cancelled, no_show]
        paymentStatus:
          type: string
          enum: [unpaid, partial, paid, refunded, failed]
        paymentMethod:
          type: string
          enum: [online_card, bank_transfer, cash, pending]
        bookingSource:
          type: ["string", "null"]
        adminNotes:
          type: ["string", "null"]
        services:
          type: array
          items:
            $ref: "#/components/schemas/BookingService"
        createdAt:
          type: string
    BookingService:
      type: object
      required: [id, serviceId, serviceName, quantity, unitPrice, subtotal]
      properties:
        id:
          type: string
        serviceId:
          type: string
        serviceName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: string
        subtotal:
          type: string
    BookingInput:
      type: object
      required: [roomId, checkIn, checkOut, guestCount, guestFullName, guestEmail, guestPhone, paymentMethod]
      properties:
        roomId:
          type: string
        checkIn:
          type: string
          format: date
        checkOut:
          type: string
          format: date
        guestCount:
          type: integer
        guestFullName:
          type: string
        guestEmail:
          type: string
        guestPhone:
          type: string
        guestNationality:
          type: ["string", "null"]
        specialRequests:
          type: ["string", "null"]
        serviceIds:
          type: array
          items:
            type: string
        paymentMethod:
          type: string
          enum: [online_card, bank_transfer, cash, pending]
        languageUsed:
          type: ["string", "null"]
...
    BookingCheck:
      type: object
      required: [roomId, checkIn, checkOut, guestCount]
      properties:
        roomId:
          type: string
        checkIn:
          type: string
          format: date
        checkOut:
          type: string
          format: date
        guestCount:
          type: integer
        serviceIds:
          type: array
          items:
            type: string
    BookingQuote:
      type: object
      required: [available, roomRatePerNight, nights, roomSubtotal, servicesSubtotal, cleaningFee, totalAmount, currency]
      properties:
        available:
          type: boolean
        roomRatePerNight:
          type: string
        nights:
          type: integer
        roomSubtotal:
          type: string
        servicesSubtotal:
          type: string
        cleaningFee:
          type: string
        totalAmount:
          type: string
        currency:
          type: string
        unavailableDates:
          type: array
          items:
            type: string
...
    ContactInput:
      type: object
      required: [fullName, email, message]
      properties:
        fullName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        message:
          type: string
        countryCode:
          type: ["string", "null"]
    ContactResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
...
    PageViewInput:
      type: object
      required: [sessionId, page]
      properties:
        sessionId:
          type: string
        page:
          type: string
        referrer:
          type: ["string", "null"]
        deviceType:
          type: ["string", "null"]
        language:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
    TrackResult:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
...